INR_Classifier/requirements.txt
azure-cognitiveservices-vision-customvision==3.0.0
azure-common==1.1.25
camera==1.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
future==0.18.2
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msrest==0.6.17
numpy==1.19.0
oauthlib==3.1.0
opencv-python==4.2.0.34
pyngrok==4.1.3
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.25.9
Werkzeug==1.0.1